requirements.txt
ortools==9.9.3963
matplotlib==3.8.4
matplotlib-inline==0.1.7
Mesa==2.3.0
Mesa-Viz-Tornado==0.1.3
networkx==3.2.1
pandas==2.2.2
tabulate==0.9.0

Flask==2.3.2
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
python-socketio==5.11.2
openpyxl==3.1.2
requests==2.32.2
urllib3==1.26.7

pyparsing==3.1.2
six==1.16.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
PyYAML==6.0.1